[week03] mission Refactor
